Email sync stoppage fix (Upstream Issue 14466) - Attempt #2
This is a second attempt at fixing this. Specifically
there was no check on the Exchange account count
so this would loop endlessly if you had no accounts
or if you had an IMAP or GMail account only.  A different
strategy is chosen per a Google employee in the upstream
issue comments.  Marc suggests just starting the service
which appears to work very well!  All checks for
configured Exchange accounts are done in the existing
code per the onStartCommand(Intent,int,int) logic.

Change-Id: Id89b7a4a0b02380a902adba9cd7d7bf53aca4cdb